Update and delete the new attribute fracturswitch and antiqua
The proposal of two attributes `fracturswitsh` and `antiqua` has proved
impracticable. Both attributes represent a specific interpretation of
the template, which cannot be done by an OCR/OLR, but only by a direct
comparison of the specific circumstances. Since the font information is
stored in the `fontFamily` attribute, this means that decisive
information is available that, if necessary, can lead in subsequent
steps of an analysis of the template to the general conclusion that this
is a change of fracture or, among other things, an antiqua passage.